+++ .rsink/constants.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+version="rsink 0.1+dev"
+config_file="config"
+profiles_directory="profiles"
+tools_directory="tools"
+log_file="log"